requirements.txt
numpy
pandas
scikit-learn
matplotlib
scipy
seaborn
tensorflow
Flask
Pillow
kagglehub[pandas-datasets]
